requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.8.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.1.0
flake8==7.2.0
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
mccabe==0.7.0
olefile==0.47
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==6.1.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
pyotp==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-factoryboy==2.7.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.2
redis==6.0.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.24.0
sentry-sdk==2.38.0
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stripe==12.5.1
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
structlog==23.2.0
django-structlog==8.1.0

# Type checking (desarrollo)
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14